+++ packages/app/scripts/genversion-android.js
@@ -1,3 +1,16 @@
+const fs = require('fs');
+const path = require('path');
+
+const version = require('../lib/version');
+const outputPath = path.resolve(
+  __dirname,
+  '..',
+  'android',
+  'src/reactnative/java/io/invertase/firebase/app',
+  'ReactNativeFirebaseVersion.java',
+);
+const template = `
+package io.invertase.firebase.app;
 /*
  * Copyright (c) 2016-present Invertase Limited & Contributors
  *
@@ -14,18 +27,10 @@
  * limitations under the License.
  *
  */
+// generated file - do not modify or commit
+public class ReactNativeFirebaseVersion {
+  public static String VERSION = "version_number";
+}
+`;
 
-export default {
-  ServerValue: {
-    TIMESTAMP: {
-      '.sv': 'timestamp',
-    },
-    increment(delta: number) {
-      return {
-        '.sv': {
-          increment: delta,
-        },
-      };
-    },
-  },
-};
+fs.writeFileSync(outputPath, template.replace('version_number', `${version}`), 'utf8');