+++ packages/auth/lib/index.d.ts
@@ -851,7 +851,7 @@ export namespace FirebaseAuthTypes {
    * ```js
    * firebase.auth().onAuthStateChanged((user) => {
    *   if (user) {
-   *     console.log('User email: ', user.email');
+   *     console.log('User email: ', user.email);
    *   }
    * });
    * ```
@@ -862,7 +862,7 @@ export namespace FirebaseAuthTypes {
    * const user = firebase.auth().currentUser;
    *
    * if (user) {
-   *  console.log('User email: ', user.email');
+   *  console.log('User email: ', user.email);
    * }
    * ```
    */