+++ packages/messaging/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.1.6](https://github.com/invertase/react-native-firebase/compare/@react-native-firebase/messaging@7.1.5...@react-native-firebase/messaging@7.1.6) (2020-06-18)
+
+### Bug Fixes
+
+- **ios:** handle nil messaging token ([#3790](https://github.com/invertase/react-native-firebase/issues/3790)) ([b2ffe11](https://github.com/invertase/react-native-firebase/commit/b2ffe115874c422414d58038d5d01e1b411c5a6c))
+
+## [7.1.5](https://github.com/invertase/react-native-firebase/compare/@react-native-firebase/messaging@7.1.4...@react-native-firebase/messaging@7.1.5) (2020-06-10)
+
+**Note:** Version bump only for package @react-native-firebase/messaging
+
+## [7.1.4](https://github.com/invertase/react-native-firebase/compare/@react-native-firebase/messaging@7.1.3...@react-native-firebase/messaging@7.1.4) (2020-06-03)
+
+**Note:** Version bump only for package @react-native-firebase/messaging
+
+## [7.1.3](https://github.com/invertase/react-native-firebase/compare/@react-native-firebase/messaging@7.1.2...@react-native-firebase/messaging@7.1.3) (2020-05-29)
+
+**Note:** Version bump only for package @react-native-firebase/messaging
+
+## [7.1.2](https://github.com/invertase/react-native-firebase/compare/@react-native-firebase/messaging@7.1.1...@react-native-firebase/messaging@7.1.2) (2020-05-29)
+
+**Note:** Version bump only for package @react-native-firebase/messaging
+
 ## [7.1.1](https://github.com/invertase/react-native-firebase/compare/@react-native-firebase/messaging@7.1.0...@react-native-firebase/messaging@7.1.1) (2020-05-28)
 
 ### Bug Fixes